New save/preserve/publish functionality; Add bixby for rubocop settings

+++ lib/hyacinth/adapters/storage_adapter/disk.rb
@@ -32,7 +32,7 @@ module Hyacinth
         end
 
         def delete(location_uri)
-          File.delete(location_uri_to_file_path(location_uri))
+          File.delete(location_uri_to_file_path(location_uri)) if File.exist?(location_uri_to_file_path(location_uri))
         end
 
         def read_impl(location_uri)
